Conflicts:
	app/controllers/meeting_contents_controller.rb
	app/helpers/meeting_contents_helper.rb
	app/models/meeting_content.rb
	app/views/meeting_contents/_form.html.erb
	config/locales/de.yml
	config/locales/en.yml
	features/meetings_locking.feature
	init.rb

+++ db/migrate/20110224180804_add_lock_to_meeting_content.rb
@@ -1,11 +1,13 @@
 class AddLockToMeetingContent < ActiveRecord::Migration
   def self.up
     add_column :meeting_contents, :locked, :boolean, :default => false
-    add_column :meeting_content_versions, :locked, :boolean, :default => nil
+    # Check for existence of the pre-journalized MeetingContentVersions table
+    add_column :meeting_content_versions, :locked, :boolean, :default => nil if table_exists? :meeting_content_versions
   end
 
   def self.down
     remove_column :meeting_contents, :locked
-    remove_column :meeting_content_versions, :locked
+    # Check for existence of the pre-journalized MeetingContentVersions table
+    remove_column :meeting_content_versions, :locked if table_exists? :meeting_content_versions
   end
 end
